+++ src/CLIMA.jl
@@ -4,12 +4,12 @@ include("Utilities/ParametersType/src/ParametersType.jl")
 include("Utilities/PlanetParameters/src/PlanetParameters.jl")
 include("Utilities/RootSolvers/src/RootSolvers.jl")
 include("Utilities/MoistThermodynamics/src/MoistThermodynamics.jl")
-include("ClimaAtmos/Parameterizations/SurfaceFluxes/src/SurfaceFluxes.jl")
-include("ClimaAtmos/Parameterizations/TurbulenceConvection/src/TurbulenceConvection.jl")
+include("Atmos/Parameterizations/SurfaceFluxes/src/SurfaceFluxes.jl")
+include("Atmos/Parameterizations/TurbulenceConvection/src/TurbulenceConvection.jl")
 include("Mesh/Topologies.jl")
 include("Mesh/Grids.jl")
 include("Arrays/MPIStateArrays.jl")
-include("ClimaAtmos/Dycore/src/CLIMAAtmosDycore.jl")
+include("Atmos/Dycore/src/AtmosDycore.jl")
 include("ODESolvers/ODESolvers.jl")
 include("ODESolvers/LowStorageRungeKuttaMethod.jl")
 include("ODESolvers/GenericCallbacks.jl")
